+++ src/conjunto/templatetags/conjunto.py
@@ -9,7 +9,7 @@ register = template.Library()
 @register.simple_tag
 def conjunto_css_scripts() -> str:
     css = [
-        # "conjunto/css/bootstrap.min.css",
+        "conjunto/css/conjunto.css",
         "conjunto/css/tabler.min.css",
         "conjunto/css/tabler-icons.css",
         "conjunto/css/dropzone.min.css",