+++ src/conjunto/__init__.py
@@ -1,3 +1,3 @@
 "Django application framework/helpers using HTMX, tabler.io, tables2, crispy-forms & more."
 
-__version__ = "0.0.13"
+__version__ = "0.0.14"