+++ CHANGELOG.md
@@ -6,8 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [0.0.14] - unreleased
-### Changed
+- fix issues with tabler https://github.com/tabler/tabler/issues/1836
 - breaking: rename required_permissions attribute to permission_required in IMenuItems
+- improve link functionality
 ## [0.0.13] - 2023-02-27
 ### Added
 - let listitem component accept subtitle slot additionally to attribute